requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
paypal-checkout-serversdk==1.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
werkzeug==3.0.1
requests==2.31.0
hmac-signature==0.1.0
cryptography==42.0.0
